--- Assets/Scripts/Game/GameManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d950a13ac6cc91b41a01e27e4952d312
