+++ Assets/Scripts/Game/GameManager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 55ca0c9dc08b5254d8c72db568c87daa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
