+++ Assets/Scripts/Game/GameManager/GameManager.ActionPanel.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: dad6b1e9d0b08e34ab5cba63c43782b4
