+++ Assets/Scripts/Game/GameManager/GameManager.Animation.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 98a75e44a91dd584b83fcfefe3d2efa4
