+++ Assets/Scripts/Game/GameManager/GameManager.UI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 730675a1fac99114a8f8e2f75072a186
