+++ Assets/Scripts/Game/GameManager/GameManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 748ad97e61cf72d49834ee6cd6c283bc
